# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * DataVault Nigeria customer wallet, bill payment, and administration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  phone: string | null;
  role: UserProfileRole;
  isActive: boolean;
  createdAt: string;
}

export interface UserProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name?: string;
  /**
     * @maxLength 30
     * @nullable
     */
  phone?: string | null;
}

export type WalletCurrency = typeof WalletCurrency[keyof typeof WalletCurrency];


export const WalletCurrency = {
  NGN: 'NGN',
} as const;

export interface Wallet {
  id: number;
  /** @minimum 0 */
  balanceKobo: number;
  currency: WalletCurrency;
  updatedAt: string;
}

export type TransactionCategory = typeof TransactionCategory[keyof typeof TransactionCategory];


export const TransactionCategory = {
  airtime: 'airtime',
  mobile_data: 'mobile_data',
  electricity: 'electricity',
  cable_tv: 'cable_tv',
  digital_subscription: 'digital_subscription',
  wallet_top_up: 'wallet_top_up',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  successful: 'successful',
  pending: 'pending',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  category: TransactionCategory;
  provider: string;
  description: string;
  accountReference: string;
  /** @minimum 0 */
  amountKobo: number;
  status: TransactionStatus;
  isTest: boolean;
  createdAt: string;
}

export type ServiceProductCategory = typeof ServiceProductCategory[keyof typeof ServiceProductCategory];


export const ServiceProductCategory = {
  airtime: 'airtime',
  mobile_data: 'mobile_data',
  electricity: 'electricity',
  cable_tv: 'cable_tv',
  digital_subscription: 'digital_subscription',
} as const;

export interface ServiceProduct {
  id: string;
  category: ServiceProductCategory;
  provider: string;
  name: string;
  description: string;
  /**
     * @minimum 0
     * @nullable
     */
  priceKobo: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  minAmountKobo: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  maxAmountKobo: number | null;
}

export interface MonthlySpend {
  month: string;
  /** @minimum 0 */
  amountKobo: number;
}

export interface DashboardSummary {
  /** @minimum 0 */
  balanceKobo: number;
  /** @minimum 0 */
  spentThisMonthKobo: number;
  /** @minimum 0 */
  transactionCount: number;
  monthlySpend: MonthlySpend[];
  recentTransactions: Transaction[];
}

export interface WalletTopUpInput {
  /**
     * @minimum 100000
     * @maximum 50000000
     */
  amountKobo: number;
}

export interface WalletTopUpResult {
  wallet: Wallet;
  transaction: Transaction;
}

export interface PurchaseInput {
  /** @minLength 1 */
  serviceId: string;
  /**
     * @minLength 3
     * @maxLength 120
     */
  accountReference: string;
  /** @minimum 100 */
  amountKobo?: number;
}

export interface Notification {
  id: number;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type AdminTransaction = Transaction & {
  userId: string;
  userEmail: string;
  userName: string;
};

export interface AdminOverview {
  /** @minimum 0 */
  totalUsers: number;
  /** @minimum 0 */
  activeUsers: number;
  /** @minimum 0 */
  transactionsToday: number;
  /** @minimum 0 */
  volumeTodayKobo: number;
  /** @minimum 0 */
  pendingTransactions: number;
  recentTransactions: AdminTransaction[];
}

export type AdminUserRole = typeof AdminUserRole[keyof typeof AdminUserRole];


export const AdminUserRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface AdminUser {
  id: string;
  email: string;
  name: string;
  /** @nullable */
  phone: string | null;
  role: AdminUserRole;
  isActive: boolean;
  /** @minimum 0 */
  balanceKobo: number;
  createdAt: string;
}

export type AdminUserUpdateRole = typeof AdminUserUpdateRole[keyof typeof AdminUserUpdateRole];


export const AdminUserUpdateRole = {
  customer: 'customer',
  admin: 'admin',
} as const;

export interface AdminUserUpdate {
  role?: AdminUserUpdateRole;
  isActive?: boolean;
}

export type ListTransactionsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListAdminTransactionsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};